moved person survey to question folder
